requirements.txt
# Core Dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
pydantic>=2.0.0

# NLP Tools
spacy>=3.7.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.7.0
